+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityConditionFieldFactory.java
@@ -48,9 +48,9 @@ import static is.codion.swing.framework.ui.component.EntityComponents.entityComp
 import static java.util.Objects.requireNonNull;
 
 /**
- * A default field factory implementation.
+ * A default field factory implementation for attributes.
  */
-public final class EntityConditionFieldFactory implements FieldFactory<Attribute<?>> {
+public final class EntityConditionFieldFactory implements FieldFactory {
 
 	private static final List<Class<?>> SUPPORTED_TYPES = Arrays.asList(
 					Character.class, String.class, Boolean.class, Short.class, Integer.class, Double.class,
@@ -58,12 +58,15 @@ public final class EntityConditionFieldFactory implements FieldFactory<Attribute
 					LocalDateTime.class, OffsetDateTime.class, Entity.class);
 
 	private final EntityComponents inputComponents;
+	private final Attribute<?> attribute;
 
 	/**
 	 * @param entityDefinition the entity definition
+	 * @param attribute the attribute
 	 */
-	public EntityConditionFieldFactory(EntityDefinition entityDefinition) {
+	public EntityConditionFieldFactory(EntityDefinition entityDefinition, Attribute<?> attribute) {
 		this.inputComponents = entityComponents(entityDefinition);
+		this.attribute = requireNonNull(attribute);
 	}
 
 	@Override
@@ -72,9 +75,9 @@ public final class EntityConditionFieldFactory implements FieldFactory<Attribute
 	}
 
 	@Override
-	public <T> JComponent createEqualField(ConditionModel<T> condition, Attribute<?> attribute) {
+	public <T> JComponent createEqualField(ConditionModel<T> condition) {
 		if (attribute instanceof ForeignKey) {
-			return createEqualForeignKeyField(condition, attribute);
+			return createEqualForeignKeyField(condition);
 		}
 
 		return inputComponents.component((Attribute<T>) attribute)
@@ -83,7 +86,7 @@ public final class EntityConditionFieldFactory implements FieldFactory<Attribute
 	}
 
 	@Override
-	public <T> Optional<JComponent> createUpperBoundField(ConditionModel<T> condition, Attribute<?> attribute) {
+	public <T> Optional<JComponent> createUpperBoundField(ConditionModel<T> condition) {
 		Class<?> columnClass = condition.valueClass();
 		if (columnClass.equals(Boolean.class) || columnClass.equals(Entity.class)) {
 			return Optional.empty();//no upper bound field required for booleans or entities
@@ -95,7 +98,7 @@ public final class EntityConditionFieldFactory implements FieldFactory<Attribute
 	}
 
 	@Override
-	public <T> Optional<JComponent> createLowerBoundField(ConditionModel<T> condition, Attribute<?> attribute) {
+	public <T> Optional<JComponent> createLowerBoundField(ConditionModel<T> condition) {
 		Class<T> columnClass = condition.valueClass();
 		if (columnClass.equals(Boolean.class) || columnClass.equals(Entity.class)) {
 			return Optional.empty();//no lower bound field required for booleans or entities
@@ -107,9 +110,9 @@ public final class EntityConditionFieldFactory implements FieldFactory<Attribute
 	}
 
 	@Override
-	public <T> JComponent createInField(ConditionModel<T> condition, Attribute<?> attribute) {
+	public <T> JComponent createInField(ConditionModel<T> condition) {
 		if (attribute instanceof ForeignKey) {
-			return createInForeignKeyField((ConditionModel<Entity>) condition, (ForeignKey) attribute);
+			return createInForeignKeyField((ConditionModel<Entity>) condition);
 		}
 
 		return listBox((ComponentValue<T, JComponent>)
@@ -117,7 +120,7 @@ public final class EntityConditionFieldFactory implements FieldFactory<Attribute
 										.buildValue(), condition.operands().in()).build();
 	}
 
-	private <T> JComponent createEqualForeignKeyField(ConditionModel<T> model, Attribute<?> attribute) {
+	private <T> JComponent createEqualForeignKeyField(ConditionModel<T> model) {
 		if (model instanceof ForeignKeyConditionModel) {
 			EntitySearchModel searchModel = ((ForeignKeyConditionModel) model).equalSearchModel();
 
@@ -135,7 +138,8 @@ public final class EntityConditionFieldFactory implements FieldFactory<Attribute
 		throw new IllegalArgumentException("Unknown foreign key condition model type: " + model);
 	}
 
-	private JComponent createInForeignKeyField(ConditionModel<Entity> model, ForeignKey foreignKey) {
+	private JComponent createInForeignKeyField(ConditionModel<Entity> model) {
+		ForeignKey foreignKey = (ForeignKey) attribute;
 		if (model instanceof ForeignKeyConditionModel) {
 			EntitySearchModel searchModel = ((ForeignKeyConditionModel) model).inSearchModel();